==> 1_data_preparation.ipynb <==
from pathlib import Path
from IPython.display import display

import numpy as np
import pandas as pd

import spotipy
import lib_spotify_app.api_adapter as api_adapter

pd.set_option('max_columns', None)

credential_fp = Path(r'private/spotify_credential.json')

I download all the saved/liked tracks from my Spotify account and their audio features:
https://developer.spotify.com/documentation/web-api/reference/tracks/get-several-audio-features/

sp = api_adapter.setup_spotipy(
    credential_fp,
    scope=['user-library-read','user-top-read'],
    cache_path=Path(r'private')
)

df = api_adapter.query_liked_songs(sp)
df = api_adapter.enrich_audiofeature(df, sp, col="track.id")

df = df.drop('index', axis=1)
df['added_at'] = pd.to_datetime(df['added_at'])

I notice that I can know when I saved/liked the song, I would like to know when I was the most active:

Cleaning of the columns for analysis, names are too complex and I will concatenate the "artists" into a list column

df.columns = df.columns.str.replace('^(track\.)(id\.)?', '')
df = df.loc[:,~df.columns.duplicated()]
df.columns.values

Concatenate the artists values into 1 column for:
* names
* id

df['artists.name'] = df.filter(regex='^artists\.\d+\.name')\
                       .apply(lambda x: x.dropna().to_list(), axis=1)
df['artists.id'] = df.filter(regex='^artists\.\d+\.id')\
                     .apply(lambda x: x.dropna().to_list(), axis=1)
df['duration_min'] = df['duration_ms'] / 60000

df.sort_values('popularity', ascending=False)[['name', 'artists.name', 'preview_url', 'external_urls.spotify', 'popularity']].head(10)

df.to_csv(Path(r'private/data.csv'), sep='\t')

We can use the LastFM app to enrich the songs features with the number of listening from the user

last_api = api_adapter.setup_lastfm(Path('private', 'lastfm_credential.json'))
top_tracks_lastfm = pd.DataFrame(
    last_api.user.get_top_tracks(period='overall', limit=500)
)
display(top_tracks_lastfm)

Issue is that LastFM and Spotify don't share the same songs ID or data management.
